+++ client/subscribe.jsx
@@ -1,3 +1,9 @@
+/* this file should not exist, subscriptions should be on a page by page or component by component basis
+ - rules for react component:
+ 1. if you use a resource, subscribe to it
+ 2. subscriptions should be coded to be as narrow as possible
+ 3. publications should be handled by controller */
+
 var handle = Meteor.subscribe('teamsAndFacilitiesForUser');
 Meteor.subscribe('users');